+++ ContentPatcher/release-notes.md
@@ -7,6 +7,7 @@
 * Patches now update immediately when you change language.
 * Fixed `EditData` patches not always updated if they use `FromFile` to load a file containing tokens.
 * Fixed patches not always updated for a `Random` token reroll.
+* Fixed field delimiter when editing `Data\Achievements`.
 * Removed support for `FromFile` in `EditData` patches in newer format versions.
 
 **Update note for mod authors:**